refactor: Clean up and optimize FloatingEdge
