+++ mikado-app/src/pages/Graph/Plaza.css
@@ -12,6 +12,11 @@
   flex-direction: column;
 }
 
+.react-flow__edge-path {
+  stroke-width: 3;
+  stroke: black;
+}
+
 main > main {
   flex-grow: 1;
   width: 100%;
